+++ katas/balancedBrackets/balancedBrackets.js
