+++ katas/camelCase/camelCase.js
