+++ katas/ensure/ensure.js
