+++ katas/formatDate/formatDate.js
