+++ katas/leftRotation/leftRotation.js
