+++ katas/prefixSet/prefixSet.js
